move arm kernel to host kernel,test=develop
