+++ lite/kernels/host/CMakeLists.txt
@@ -121,6 +121,8 @@ add_kernel(roll_compute Host extra SRCS roll_compute.cc)
 add_kernel(set_value Host extra SRCS set_value_compute.cc)
 add_kernel(share_data_compute_host Host extra SRCS share_data_compute.cc)
 add_kernel(round_compute_host Host extra SRCS round_compute.cc)
+add_kernel(temporal_shift_compute_host Host extra SRCS temporal_shift_compute.cc)
+
 
 if(LITE_BUILD_EXTRA AND LITE_WITH_x86)
   lite_cc_test(test_where_index_compute_host SRCS where_index_compute.cc)
